% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allowed_values_categorical.R
\name{check_allowed_values_categorical}
\alias{check_allowed_values_categorical}
\title{Check allowed values for categorical variables}
\usage{
check_allowed_values_categorical(df, allowed_values)
}
\arguments{
\item{df}{Data frame of categorical variables to check values of}

\item{allowed_values}{A list where names are colnames and items
are vectors of allowed values for the column}
}
\value{
A list where names are colnames and items are vectors of
unexpected values observed in the data from df.
If no unexpected values are observed, return NULL.
}
\description{
Does not account for allowed values that are not observed.
}
