% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allowed_values_numeric.R
\name{check_allowed_values_numeric}
\alias{check_allowed_values_numeric}
\title{Check allowed values for numeric variables}
\usage{
check_allowed_values_numeric(df, allowed_min, allowed_max)
}
\arguments{
\item{df}{Data frame containing numeric variables to check}

\item{allowed_min}{Named vector where names are colnames and values are
the allowed minimum value for the variable.}

\item{allowed_max}{Named vector where names are colnames and values are
the allowed maximum value for the variable.

It is not required that allowed_min and allowed_max have the same set of
colnames specified to check (their names don't have to be the same set
of colnames).}
}
\value{
NULL if no unexpected values are found. Otherwise, a dataframe
specifying count below the min or above the max for each variable.
}
\description{
Check allowed values for numeric variables
}
