% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_types.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data types match expected}
\usage{
check_data_types(df, expected_types)
}
\arguments{
\item{df}{Data frame to check column types for}

\item{expected_types}{Named vector with colname = datatype pairs;
datatype should be the result of typeof() on the column}
}
\value{
NULL if all data types match; data frame describing mismatches if not
}
\description{
Check data types match expected
}
