% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_strings.R
\name{check_empty_strings}
\alias{check_empty_strings}
\title{Check varnames in df for empty character strings}
\usage{
check_empty_strings(df, varnames = NULL)
}
\arguments{
\item{df}{Data frame containing variables to check for empty string values.}

\item{varnames}{Character vector of variable names from df
to check for empty string values values.}
}
\value{
Named vector where names are varnames and values are
the count of empty string values for the variable.
}
\description{
Check varnames in df for empty character strings
}
