% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_values.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check varnames in df for missing values}
\usage{
check_missing_values(df, varnames = NULL)
}
\arguments{
\item{df}{Data frame containing variables to check for missing values.}

\item{varnames}{Character vector of variable names from df
to check for missing values.}
}
\value{
Named vector where names are varnames and values are
the count of missing values for the variable.
}
\description{
Check varnames in df for missing values
}
