% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate records from a dataframe}
\usage{
remove_duplicates(df, dup_idx)
}
\arguments{
\item{df}{Data frame to remove duplicates from}

\item{dup_idx}{Indexes of records (rows) to remove from the df.}
}
\value{
df with the specified rows removed.
}
\description{
Remove duplicate records from a dataframe
}
