% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cleaned_data.R
\name{save_cleaned_data}
\alias{save_cleaned_data}
\title{Save cleaned data to file(s) named based on the original file name}
\usage{
save_cleaned_data(
  clean_df,
  original_path,
  save_rds = TRUE,
  save_csv = TRUE,
  cleaned_modifier = "_cleaned"
)
}
\arguments{
\item{clean_df}{Data frame containing cleaned data to save}

\item{original_path}{Path to original data file that has been cleaned}

\item{save_rds}{bool; whether to save data to an RDS file}

\item{save_csv}{book; whether to save data to a CSV file}
}
\value{
Returns NULL; saves data to the current working dir with a filename
based on original_path.
}
\description{
Save cleaned data to file(s) named based on the original file name
}
